tests/requirements.txt
pyrogram==2.0.104
#telethon==1.28.5
#tgcrypto==1.2.5
#uvloop==0.17.0
#pytest==7.3.1
python-decouple==3.8
tgintegration==1.2.0
git+https://github.com/GeiserX/tgintegration.git==2.0.0
python-dotenv==1.0.0